+++ app/src/main/java/com/luckyzyx/luckytool/hook/statusbar/StatusBarClock.kt
@@ -4,6 +4,7 @@ import com.highcapable.yukihookapi.hook.entity.YukiBaseHooker
 import com.luckyzyx.luckytool.hook.scope.systemui.StatusBarClock
 import com.luckyzyx.luckytool.utils.tools.XposedPrefs
 
+
 object StatusBarClock : YukiBaseHooker() {
     override fun onHook() {
         loadApp("com.android.systemui"){